fix compile warnings
git-svn-id: https://simulationcraft.googlecode.com/svn/trunk@4161 3b4652a1-8050-0410-ac47-3d40261b0f8b

+++ engine/sc_chart.cpp
@@ -790,7 +790,7 @@ const char* chart_t::scale_factors( std::string& s,
   {
     if( p -> scales_with[ i ] )
     {
-      float s = p -> scaling.get_stat( i );
+      double s = p -> scaling.get_stat( i );
       
       if ( s > 0 ) scaling_stats.push_back( i );
     }
